fix(EditableHtml): wrap the Editor with a component that checks markup against value
