+++ packages/editable-html/demo/rte-demo.jsx
@@ -12,7 +12,8 @@ const puppySrc = 'http://cdn2-www.dogtime.com/assets/uploads/gallery/30-impossib
  */
 // const html = `<div><div>hi</div><img src="${puppySrc}"></img></div>`;
 // const html = `<span data-mathjax="">\\frac{1}{2}</span>`;
-const html = `<ul><li><span>apple<span></li></ul>`;
+// const html = `<ul><li><span>apple<span></li></ul>`;
+const html = `hi`;
 
 
 // const j = { "kind": "value", "document": { "kind": "document", "data": {}, "nodes": [{ "kind": "block", "type": "div", "nodes": [{ "kind": "text", "leaves": [{ "kind": "leaf", "text": "a" }] }, { "kind": "block", "type": "image", "isVoid": true, "nodes": [], "data": { "src": "http://cdn2-www.dogtime.com/assets/uploads/gallery/30-impossibly-cute-puppies/impossibly-cute-puppy-8.jpg", "width": null, "height": null } }] }] } }
@@ -133,14 +134,12 @@ export default class RteDemo extends React.Component {
         imageSupport={imageSupport}
         activePlugins={['bold', 'bulleted-list', 'numbered-list']}
         onBlur={this.onBlur}
-        width="400px"
-        height="300px"
       />
       <input type="file" hidden ref={r => this.fileInput = r}></input>
       <br />
       <br />
       <h4>markup</h4>
-      <pre>{markup}</pre>
+      <pre className="prettyprint">{markup}</pre>
     </div>);
   }
 }
