+++ packages/editable-html/package.json
@@ -20,6 +20,7 @@
     "react-jss": "^7.0.2",
     "react-portal": "^3.1.0",
     "slate": "^0.33.0",
+    "slate-prop-types": "^0.4.23",
     "slate-html-serializer": "^0.5.6",
     "slate-react": "^0.12.0",
     "slate-edit-list": "^0.11.2"
@@ -47,6 +48,8 @@
     "react-test-renderer": "^15.6.1",
     "sinon": "^4.1.3",
     "style-loader": "^0.18.2",
+    "slate-typings": "^0.0.3",
+    "slate-prop-types-typings": "^0.0.3",
     "url-loader": "^0.5.9",
     "webpack": "^3.5.4",
     "webpack-dashboard": "^1.0.0-2",
