+++ packages/editable-html/src/plugins/index.jsx
@@ -5,14 +5,12 @@ import NumberedListIcon from 'material-ui-icons/FormatListNumbered';
 import ImagePlugin from './image';
 import Italic from 'material-ui-icons/FormatItalic';
 import MathPlugin from './math';
-import PropTypes from 'prop-types';
 import React from 'react';
 import Strikethrough from 'material-ui-icons/FormatStrikethrough';
 import ToolbarPlugin from './toolbar';
 import Underline from 'material-ui-icons/FormatUnderlined';
 import compact from 'lodash/compact';
 import debug from 'debug';
-import BulletedList from './bulleted-list';
 import List from './list';
 
 const log = debug('editable-html:plugins');
@@ -78,10 +76,8 @@ export const buildPlugins = (activePlugins, opts) => {
     addIf('underline', MarkHotkey({ key: 'u', type: 'underline', icon: <Underline />, tag: 'u' })),
     addIf('image', opts.image && opts.image.onDelete && ImagePlugin(opts.image)),
     addIf('math', MathPlugin(opts.math)),
-    //addIf('bulleted-list', BulletedList({ key: 'l', type: 'bulleted-list', icon: <BulletedListIcon /> })),
     addIf('bulleted-list', List({ key: 'l', type: 'ul_list', icon: <BulletedListIcon /> })),
     addIf('numbered-list', List({ key: 'n', type: 'ol_list', icon: <NumberedListIcon /> })),
-    ToolbarPlugin(opts.toolbar),
-
+    ToolbarPlugin(opts.toolbar)
   ]);
 }
